Updating start.sh: Support to git + volume: iwaitnotify instead of poll
This commit adds support for iwaitnotify. That is, wait on files to be
added to DATA_PATH. That allows us to use this image in a Docker-Compose
environment that can wait on files to be built. For instance, building
expensive .apk files for Android projects to be pushed to S3.